// src/api/models/XsPatches.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type XsPatches = {
    uuid?: string;
    url?: string;
    upgrade?: boolean;
    requirements?: Array<string>;
    paid?: boolean;
    id?: string;
    name: string;
    guidances: string;
    documentationUrl?: string;
    description: string;
    date: string;
    conflicts?: Array<string>;
};